Load dynamic library using two alternative methods
The first method is borrowed from #240 / GeoDjango, which is to use an
environment variable GEOS_LIBRARY_PATH to point directly to a dynamic library.
If this seems like a hack, it can be removed. However, it could be useful
for folks that don't have GEOS_CONFIG, such as Windows, and want to use a
different DLL file.

The second method is to use geos-config from the PATH, or can be explicity set
using a GEOS_CONFIG flag. It attempts a few paths to libraries from the output
from `geos-config --clibs`. If used in combination with GEOS_LIBRARY_PATH,
the two library versions are checked by setup.py.

After these attempts, the old code is still in place (with very minor changes).

An unused `free` object is still a TODO.