An improved method of not deleting EMPTY
During python shutdown all modules are destroyed in an arbitrary order.
So the shapely modules may be destroyed before modules holding objects
created using shapely classes. This creates a race between the __del__ on
the object and removing EMPTY.

This change removes the need to a module level object to prevent the
data pointed to by EMPTY being destroyed by an empty shapely object