Build manylinux1 wheels
This patch uses Docker and [auditwheel][] to build [manylinux1][] wheels
for Python 2.6/3.3+. These wheels bundle GEOS into the wheel, similar to
the OS X wheels, so Shapely can be `pip install`d on most Linuxes
without GEOS or even a compiler.

All of the wheels are built with one step: `./build-linux-wheels.sh`.
This will build a Docker image based on the official manylinux images
containing GEOS and will build all of the wheels and output them in
`wheels/`. Note that this only builds x86_64 wheels, but it could be
easily extended to building i686 wheels using the
quay.io/pypa/manylinux1_i686 Docker image.

[auditwheel]: https://github.com/pypa/auditwheel
[manylinux1]: https://github.com/pypa/manylinux