Add support for reading/passing along CRS info
If a CRS is present on the context object when using `shape` or `asShape`, let's just read in the CRS for them.
This also adds the CRS to the `__geo_interface__` property (if present).
There were also some code style / import fixes and some other pep8 stuff done to some of the files along the way.