Fixing bug in ops.substring
If `min_dist` is equal to 0 (i.e. the first coordinate) then the loop immediately breaks.

This means that internal coordinates are skipped because only `(start_point.x, start_point.y)` and `(end_point.x, end_point.y)` are written to the new `vertex_list`.

By changing the less than to a less-than-equal sign the loop behaves as intended.